requirements.txt
pandas
opencv-python
omegaconf
einops
lpips
timm==0.9.16
pre-commit
ipykernel
pycocotools
pandas
PyWavelets
scikit-image
pandas
tqdm
pytorch_msssim
tensorboard
calflops
transformers
safetensors
pycocotools
opencv-python
tqdm
timm
scipy
ffmpeg-python
av
